support going to and from "Home" in New Reddit
When user of New Reddit clicks on "Home" from some subreddit URL, the
on-the-fly loading of New Reddit will leave the subreddit's icon in the
tab icon, which is undesireable.  Similarly, when a user of New Reddit
navigates to some subreddit from the home page (aka front page), then
the userscript won't be loaded because of the @match rules.

To deal with navigating to and from "Home" page of New Reddit, @match
all URLs of reddit, and not only those that have `/r/` in them.  Reset
to the default icon whenever absence of subreddit name in the URL is
detected.  This will also cover going from a subreddit to preferences,
a user's page, and other similar non-subreddit pages.  While we're here,
reset the icon to default also when a subreddit doesn't have its own
icon, so that icon of a previous subreddit doesn't linger.

Since now not finding a subreddit name would be a more normal, more
often occurring operation, downgrade the log level from `error` to
`warn` for these cases.