exclude reddit.com/account/*
New Reddit UI includes an <iframe> with some tracking, experiments, or
whatever.  The URL of the <iframe> starts with:

    https://www.reddit.com/account/sso/one_tap/?experiment...

Tampermonkey runs the script in it, but the script obviously fails.

Exclude URLs with paths `/account/*` to avoid the script being run
unnecessarily.  To make investigation of such issues easier, include the
URL of page in the main error message.

Note that @exclude is not affected by the Manifest V3, which broke meta
tag @include.  Documentation of @exclude:

    https://www.tampermonkey.net/documentation.php?locale=en#meta:exclude