update more <link> tags
I haven't figured out yet exactly which configurations (OS, browser) are
affected, but as per [1] there are situations in which Chrome 109
doesn't get the new favicons with the error in the log:

    Couldn't find favicon elements.

To attempt fix this bug, expand the selector for <link> tags to include

    <link rel="icon shortcut" ...>
    <link rel="apple-touch-icon" ...>

in the update done by the script.

To make investigation easier, add more log messages.

[1] https://greasyfork.org/en/scripts/460086-subreddit-tab-icons/discussions/198787

Repoted-by: Cool722